Move some joins to view; add initialization
1. Move attribution & activation joins to the view. This lets
   us immediately access updates to those tables, rather than
   re-materializing this table on changes there.
2. Add the capability to init from a query file. This uses
   an `is_init` jinja function, which is only set to True
   when run from `bqetl query initialize`.

+++ bigquery_etl/cli/query.py
@@ -811,6 +811,7 @@ def _run_query(
     destination_table,
     dataset_id,
     query_arguments,
+    addl_templates: dict,
 ):
     """Run a query."""
     if dataset_id is not None:
@@ -881,6 +882,7 @@ def _run_query(
                     template_folder=str(query_file.parent),
                     templates_dir="",
                     format=False,
+                    **addl_templates,
                 )
             )
             query_stream.seek(0)
@@ -1189,34 +1191,64 @@ def initialize(name, sql_dir, project_id, dry_run):
         sys.exit(1)
 
     for query_file in query_files:
-        init_files = Path(query_file.parent).rglob("init.sql")
-        client = bigquery.Client()
+        sql_content = query_file.read_text()
 
-        for init_file in init_files:
-            project = init_file.parent.parent.parent.name
+        # Enable init from query.sql files
+        # First deploys the schema, then runs the init
+        # This does not currently verify the accuracy of the schema
+        if "is_init()" in sql_content:
+            project = query_file.parent.parent.parent.name
+            dataset = query_file.parent.parent.name
+            destination_table = query_file.parent.name
+            Schema.from_schema_file(query_file.parent / SCHEMA_FILE).deploy(
+                f"{project}.{dataset}.{destination_table}"
+            )
+            arguments = [
+                "query",
+                "--use_legacy_sql=false",
+                "--replace",
+                "--format=none",
+            ]
+            _run_query(
+                query_files=[query_file],
+                project_id=project,
+                public_project_id=None,
+                destination_table=destination_table,
+                dataset_id=dataset,
+                query_arguments=arguments,
+                addl_templates={
+                    "is_init": lambda: True,
+                },
+            )
+        else:
+            init_files = Path(query_file.parent).rglob("init.sql")
+            client = bigquery.Client()
 
-            with open(init_file) as init_file_stream:
-                init_sql = init_file_stream.read()
-                dataset = Path(init_file).parent.parent.name
-                job_config = bigquery.QueryJobConfig(
-                    dry_run=dry_run,
-                    default_dataset=f"{project}.{dataset}",
-                )
+            for init_file in init_files:
+                project = init_file.parent.parent.parent.name
 
-                if "CREATE MATERIALIZED VIEW" in init_sql:
-                    click.echo(f"Create materialized view for {init_file}")
-                    # existing materialized view have to be deleted before re-creation
-                    view_name = query_file.parent.name
-                    client.delete_table(
-                        f"{project}.{dataset}.{view_name}", not_found_ok=True
+                with open(init_file) as init_file_stream:
+                    init_sql = init_file_stream.read()
+                    dataset = Path(init_file).parent.parent.name
+                    job_config = bigquery.QueryJobConfig(
+                        dry_run=dry_run,
+                        default_dataset=f"{project}.{dataset}",
                     )
-                else:
-                    click.echo(f"Create destination table for {init_file}")
 
-                job = client.query(init_sql, job_config=job_config)
+                    if "CREATE MATERIALIZED VIEW" in init_sql:
+                        click.echo(f"Create materialized view for {init_file}")
+                        # existing materialized view have to be deleted before re-creation
+                        view_name = query_file.parent.name
+                        client.delete_table(
+                            f"{project}.{dataset}.{view_name}", not_found_ok=True
+                        )
+                    else:
+                        click.echo(f"Create destination table for {init_file}")
 
-                if not dry_run:
-                    job.result()
+                    job = client.query(init_sql, job_config=job_config)
+
+                    if not dry_run:
+                        job.result()
 
 
 @query.command(
@@ -1534,7 +1566,12 @@ def _update_query_schema(
                 existing_schema.to_yaml_file(existing_schema_path)
 
     # replace temporary table references
-    sql_content = query_file_path.read_text()
+    sql_content = render_template(
+        query_file_path.name,
+        template_folder=str(query_file_path.parent),
+        templates_dir="",
+        format=False,
+    )
 
     for orig_table, tmp_table in tmp_tables.items():
         table_parts = orig_table.split(".")
